# Add __slots__ to Api class